requirements.txt
discord.py==2.3.2
anthropic>=0.40.0
notion-client>=2.2.1
